requirements.txt
pytest==8.3.3
requests==2.32.3
jsonschema==4.23.0
allure-python-commons~=2.13.5
allure-pytest
python-dotenv==0.20.0